Sort test cases by job type
